requirements.txt
pyside6
requests
python-socketio
websocket-client